services/ingestion/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
kafka-python==2.0.2
psycopg2-binary